deploy

+++ controller/user/logout.js
@@ -2,8 +2,16 @@ const models = require("../../models");
 const { ref } = models;
 
 module.exports = {
-  get: async (req, res) => {
+  post: async (req, res) => {
+if(req.cookies.refreshToken){
     await ref.destroy({ where: { hashed: req.cookies.refreshToken } });
-    res.clearCookie("refreshToken").send({ message: "logOut" });
+}
+    res.clearCookie("refreshToken", {
+          domain: "joinus.fun",
+          path: "/",
+          sameSite: "none",
+          httpOnly: true,
+          secure: true,
+        }).send({ message: "logOut" });
   },
 };
